split proximity in submodules.
